Use pybind11 to create the Python binding
Experimental branch to investigate how
difficult it is to move away from SIP
and create the Python binding with pybind11

A stand-alone SIP project can't handle C++17
standards very easily. This is because the
SIP API (written in C) is compiled as part
of a stand-alone project. We either need
to define it as a SIP package project,
write our own Builder or use it with the
no-compile argument and use CMake to link
and build. These methods are frustrating,
time-consuming and require maintenance
from our side.

Pybind11 on the other-hand seems to have
an easy API, there is a Conan package,
C++17 and higher is supported out of
the box and it is written in C++, so
IDE functionality works out of the box.

This branch is a quick investigation in
the effort rquired to switch to Pybind11